+++ VK/VK/NetworkService/Models/Friends.swift
@@ -5,24 +5,24 @@ import Foundation
 import RealmSwift
 
 /// Друзья
-final class ResponseWithFriends: Codable {
+final class ResponseWithFriends: Decodable {
     let response: Friends
 
     /// Ответ с количеством друзей и объектами "Друг"
-    class Friends: Codable {
+    class Friends: Decodable {
         let count: Int
         let items: [Friend]
     }
 }
 
 /// Место проживания
-final class City: Object, Codable {
+final class City: Object, Decodable {
     @objc dynamic var id: Int
     @objc dynamic var title: String
 }
 
 /// Друг
-final class Friend: Object, Codable {
+final class Friend: Object, Decodable {
     @objc dynamic var id: Int
     @objc dynamic var city: City?
     @objc dynamic var photo: String?
